A little tidying and bug fixing and delete name in inspect database now behaves a little better
git-svn-id: http://192.168.1.10/repos/toto@2448 b53939d4-f08c-4181-96d4-b201aa9c61c4

+++ Toto/trunk/src/com/azquo/spreadsheet/controller/JstreeController.java
@@ -151,9 +151,11 @@ public class JstreeController {
                                     .getNameDetailsJson(loggedInUser.getDataAccessToken(), currentNode.nameId));
                         }
                         if ("delete_node".equals(op)) { // still used?
-                            RMIClient.getServerInterface(loggedInUser.getDatabaseServer().getIp())
-                                    .deleteNode(loggedInUser.getDataAccessToken(), currentNode.nameId);
-                            result = "true";
+                            if (loggedInUser.getUser().isAdministrator()){
+                                RMIClient.getServerInterface(loggedInUser.getDatabaseServer().getIp())
+                                        .deleteNode(loggedInUser.getDataAccessToken(), currentNode.nameId);
+                                result = "true";
+                            }
                         }
                     }
                     if ("create_node".equals(op)) { // moved outside, it can operate with a null current node, adding to root
